rename: teams -> join
팀 전체 리스트 상수 -> 회원가입 시 필요한 상수 담는 파일로 변경
